+++ examples/less/src/main.tsx
@@ -5,6 +5,9 @@ import './main.css';
 export function Main() {
   return (
     <>
+      <div>
+        <div className="dep">less</div>
+      </div>
       <Welcome />
     </>
   );