+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/core
 
+## 1.0.8
+
+### Patch Changes
+
+- cb7df71f: Support alias resolve and url rebase for import.meta.glob, sass and less plugins
+
 ## 1.0.7
 
 ### Patch Changes